feat: implement AuthenticatedHeader and GuestHeader components for user navigation
